requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
motor==3.6.0
pymongo==4.9.2
PyJWT==2.9.0
passlib==1.7.4
python-dotenv==1.0.1
